+++ wsb-jpa1/src/main/java/com/capgemini/wsb/persistence/dao/Dao.java
@@ -21,7 +21,7 @@ public interface Dao<T, K extends Serializable> {
 
     void deleteAll();
 
-    long count();
+    Integer count();
 
     boolean exists(K id);
 }
