+++ wsb-jpa1/src/main/java/com/capgemini/wsb/persistence/dao/PatientDao.java
@@ -5,15 +5,15 @@ import com.capgemini.wsb.persistence.entity.VisitEntity;
 
 import java.util.List;
 
-public interface PatientDao extends Dao<PatientEntity, Long>
+public interface PatientDao extends Dao<PatientEntity, Integer>
 {
     List<PatientEntity> findByFirstName(String firstName);
 
     List<PatientEntity> findByLastName(String lastName);
 
-    List<VisitEntity> findVisitsByPatientId(Long id);
+    List<VisitEntity> findVisitsByPatientId(Integer id);
 
-    List<PatientEntity> findPatientsWithMoreVisitsThan(Long visitsNum);
+    List<PatientEntity> findPatientsWithMoreVisitsThan(Integer visitsNum);
 
     List<PatientEntity> findByEmploymentType(Boolean employmentType);
 }
