+++ wsb-jpa1/src/main/java/com/capgemini/wsb/rest/exception/EntityNotFoundException.java
@@ -3,7 +3,7 @@ package com.capgemini.wsb.rest.exception;
 public class EntityNotFoundException extends RuntimeException
 {
 
-    public EntityNotFoundException(Long id)
+    public EntityNotFoundException(Integer id)
     {
         super("Could not find entity of id " + id);
     }
