+++ wsb-jpa1/src/main/java/com/capgemini/wsb/service/DoctorService.java
@@ -4,5 +4,5 @@ import com.capgemini.wsb.dto.DoctorTO;
 
 public interface DoctorService
 {
-    public DoctorTO findById(final Long id);
+    public DoctorTO findById(final Integer id);
 }
